+++ CHANGELOG.md
@@ -15,7 +15,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :rocket: New Feature
 
-* Added a new Webpack plugin `SimpleProgressWebpackPlugin` for viewing current build status `build/webpack`
+* Added a new Webpack plugin `SimpleProgressWebpackPlugin` to view the current build status `build/webpack`
 
 ## v3.9.1 (2021-11-09)